fix logging support
