+++ docker-compose.yaml
@@ -56,13 +56,12 @@ services:
       - --log-level=debug
       - --server
       - --config-file=/opa.yaml
-      - --log-format=json-pretty
       - --diagnostic-addr=0.0.0.0:8282
       - --set=plugins.envoy_ext_authz_grpc.addr=:9002
-      - --set=decision_logs.console=true
       - /etc/rego/common.rego
       - /etc/rego/context_broker_policy.rego
       - /etc/rego/anubis_management_api_policy.rego
+      #- /etc/rego/audit.rego
     networks:
       - envoymesh
